// controls/richtexteditor/blazor-script/editor-scripts/common/interface.ts
export interface IImageCommandsArgs {
    /** Specifies the `src` attribute of the image. */
    url?: string
    /** Represents the current selection instance. */
    selection?: NodeSelection
    /** Specifies the minimum, maximum, and actual width of the image. */
    width?: { minWidth?: string | number; maxWidth?: string | number; width?: string | number }
    /** Specifies the minimum, maximum, and actual height of the image. */
    height?: { minHeight?: string | number; maxHeight?: string | number; height?: string | number }
    /** Describes the alternate text attribute for the image. */
    altText?: string
    /** Defines the CSS class names to be applied to the image. */
    cssClass?: string
    /** Refers to the image element that is to be edited. */
    selectParent?: Node[]
}
// ...
export interface ITouchData {
    prevClientX?: number
    prevClientY?: number
    clientX?: number
    clientY?: number
}
// ...
export interface NotifyArgs {
    module?: string
    args?: KeyboardEvent | MouseEvent | ClickEventArgs | ClipboardEvent | TouchEvent
    cancel?: boolean
    requestType?: string
    enable?: boolean
    properties?: object
    selection?: NodeSelection
    link?: HTMLInputElement
    selectNode?: Node[]
    selectParent?: Node[]
    url?: string
    text?: string
    isWordPaste?: boolean
    title?: string
    target?: string
    member?: string
    /** Specifies the name of the notifier handling the event. */
    name?: string
    /** Represents the range of text selection involved in the notification. */
    range?: Range
    /** Describes the action associated with the notification event. */
    action?: string
    callBack?(args?: string | IImageCommandsArgs, cropImageData?:
    { [key: string]: string | boolean | number }[], pasteTableSource?: string): void
    file?: Blob
    insertElement?: Element
    touchData?: ITouchData
    allowedStylePropertiesArray?: string[]
    isPlainPaste?: boolean
    formatPainterSettings?: FormatPainterSettingsModel
    ariaLabel?: string
    /**
     * Defines the source of the Table content.
     *
     * @private
     */
    pasteTableSource?: string
}